Crate subsocial_runtime

The Subsocial Node runtime.

Modules

api

constants

opaque
Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

construct_runtime
Construct a runtime, with the given name and the given modules.

parameter_types
Create new implementations of the Get trait.

Structs

AvailableBlockRatio

BaseFilter

BlockExecutionWeight
Importing a block with 0 txs takes ~5 ms

BlockHashCount

CreateCommentActionWeight

DealWithFees

DownvoteCommentActionWeight

DownvotePostActionWeight

ExistentialDeposit

ExtrinsicBaseWeight
Executing 10,000 System remarks (no-op) txs takes ~1.26 seconds -> ~125 µs per tx

FollowAccountActionWeight

FollowSpaceActionWeight

GenesisConfig

HandleDeposit

IdentityFee
Implementor of WeightToFeePolynomial that maps one unit of weight to one unit of fee.

MaxCommentDepth

MaxHandleLen

MaxLocks

MaxScheduledPerBlock

MaxUsersToProcessPerDeleteRole

MaximumBlockLength

MaximumBlockWeight
We allow for 2 seconds of compute with a 6 second average block time.

MaximumExtrinsicWeight
Assume 10% of weight for average on_initialize calls.

MaximumSchedulerWeight

MinHandleLen

MinimumPeriod

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill
A fixed point representation of a number in the range [0, 1].

Permill
A fixed point representation of a number in the range [0, 1].

RocksDbWeight
By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

ShareCommentActionWeight

SharePostActionWeight

TransactionByteFee

UpvoteCommentActionWeight

UpvotePostActionWeight

Version

Enums

BalancesCall
Dispatchable calls.

Call

Event

OriginCaller

TimestampCall
Dispatchable calls.

Constants

VERSION
This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage
Complex storage builder stuff.

Currency
Abstraction over a fungible assets system.

Filter
Simple trait for providing a filter over a reference to some type.

Imbalance
A trait for a not-quite Linear Type that tracks an imbalance.

KeyOwnerProofSystem
Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

OnUnbalanced
Handler for when some currency “account” decreased in balance for some reason.

Randomness

StorageValue
A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version
The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address

Aura

AuraConfig

Balance
Balance of an account.

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

BlockNumber
An index to a block.

CheckedExtrinsic
Extrinsic type that has already been checked.

DigestItem
Digest item type.

Executive
Executive: handles dispatch to the various modules.

Faucets

Grandpa

GrandpaConfig

Hash
A hash of some data used by the chain.

Header
Block header type as expected by this runtime.

Index
Index of a transaction in the chain.

Moment
Type used for expressing timestamp.

Permissions

PostHistory

Posts

ProfileFollows

ProfileHistory

Profiles

RandomnessCollectiveFlip

Reactions

Roles

Scheduler

Scores

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SpaceFollows

SpaceHistory

SpaceOwnership

Spaces

SpacesConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

Utils

UtilsConfig

Weight
Numeric range of a transaction weight.